feat: 광고 체험단 상품글 등록 화면 - 참여 정보 컴포넌트 구현, 하단 버튼 및 캠페인 등록 버튼 생성 및 연결 (#83)
