--- src/app/advertiser/ads-list/register/styles/CampaignRegisterForm.module.css
@@ -1,426 +1,0 @@
-.container {
-  padding: 32px 144px;
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-.formContainer {
-  display: flex;
-  gap: 32px;
-  margin-bottom: 32px;
-}
-
-.progressSection {
-  width: 300px;
-  flex-shrink: 0;
-}
-
-.progressHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 16px;
-}
-
-.progressHeader h3 {
-  font-size: 14px;
-  font-weight: 400;
-  color: #4B5563;
-  margin: 0;
-}
-
-.progressHeader span {
-  font-size: 14px;
-  font-weight: 400;
-  color: #4B5563;
-}
-
-.progressBar {
-  width: 100%;
-  height: 8px;
-  background-color: #E5E7EB;
-  border-radius: 9999px;
-  margin-bottom: 32px;
-  overflow: hidden;
-}
-
-.progressFill {
-  height: 100%;
-  background-color: #F5A623;
-  border-radius: 9999px;
-  transition: width 0.3s ease;
-}
-
-.stepsContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.stepItem {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.stepCircle {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 12px;
-  transition: all 0.3s ease;
-}
-
-.stepCircle.completed {
-  font-weight: bold;
-}
-
-.stepName {
-  font-size: 12px;
-  color: #6B7280;
-  font-weight: 400;
-}
-
-.form {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.formSection {
-  background: white;
-  border: 1px solid #E5E7EB;
-  border-radius: 12px;
-  padding: 25px;
-  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.05);
-}
-
-.label {
-  display: block;
-  font-size: 18px;
-  font-weight: 400;
-  color: #594A3E;
-  margin-bottom: 16px;
-}
-
-.required {
-  color: #EF4444;
-  margin-left: 4px;
-}
-
-.input {
-  width: 100%;
-  padding: 16px;
-  border: 1px solid #D1D5DB;
-  border-radius: 8px;
-  font-size: 16px;
-  font-weight: 400;
-  color: #594A3E;
-  background: white;
-  box-sizing: border-box;
-}
-
-.input:focus {
-  outline: none;
-  border-color: #F5A623;
-}
-
-.input::placeholder {
-  color: #ADAEBC;
-}
-
-.textarea {
-  width: 100%;
-  padding: 16px;
-  border: 1px solid #D1D5DB;
-  border-radius: 8px;
-  font-size: 16px;
-  font-weight: 400;
-  color: #594A3E;
-  background: white;
-  resize: vertical;
-  min-height: 120px;
-  box-sizing: border-box;
-  font-family: inherit;
-}
-
-.textarea:focus {
-  outline: none;
-  border-color: #F5A623;
-}
-
-.textarea::placeholder {
-  color: #ADAEBC;
-}
-
-.imageUploadArea {
-  width: 100%;
-  min-height: 216px;
-  border: 2px dashed #D1D5DB;
-  border-radius: 8px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: white;
-}
-
-.uploadPlaceholder {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-  text-align: center;
-}
-
-.uploadIcon {
-  margin-bottom: 8px;
-}
-
-.uploadPlaceholder p {
-  font-size: 16px;
-  font-weight: 400;
-  color: #4B5563;
-  margin: 0;
-}
-
-.uploadPlaceholder span {
-  font-size: 14px;
-  font-weight: 400;
-  color: #9CA3AF;
-}
-
-.uploadButton {
-  background: #F5A623;
-  color: white;
-  border: none;
-  border-radius: 8px;
-  padding: 10px 24px;
-  font-size: 16px;
-  font-weight: 400;
-  cursor: pointer;
-  transition: background-color 0.2s;
-  margin-top: 8px;
-}
-
-.uploadButton:hover {
-  background: #E0941A;
-}
-
-.imagePreview {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.arrayFieldContainer {
-  display: flex;
-  gap: 12px;
-  margin-bottom: 12px;
-  align-items: center;
-}
-
-.arrayFieldContainer .input {
-  flex: 1;
-}
-
-.removeButton {
-  width: 45px;
-  height: 45px;
-  background: #EFF6FF;
-  border: none;
-  border-radius: 50%;
-  color: #76AEF7;
-  font-size: 16px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.2s;
-  flex-shrink: 0;
-}
-
-.removeButton:hover {
-  background: #DBEAFE;
-  color: #3B82F6;
-}
-
-.addButton {
-  background: none;
-  border: none;
-  color: #F5A623;
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  padding: 8px 0;
-  transition: color 0.2s;
-}
-
-.addButton:hover {
-  color: #E0941A;
-}
-
-.dateSection {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 24px;
-}
-
-.dateField {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.dateLabel {
-  font-size: 15px;
-  font-weight: 400;
-  color: #594A3E;
-}
-
-.dateInputs {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.dateInputs span {
-  color: #6B7280;
-  font-size: 14px;
-}
-
-.datePicker {
-  flex: 1;
-  padding: 16px;
-  border: 1px solid #D1D5DB;
-  border-radius: 8px;
-  font-size: 16px;
-  font-weight: 400;
-  color: #594A3E;
-  background: white;
-  cursor: pointer;
-}
-
-.datePicker:focus {
-  outline: none;
-  border-color: #F5A623;
-}
-
-/* DatePicker 커스텀 스타일 */
-:global(.react-datepicker-wrapper) {
-  flex: 1;
-}
-
-:global(.react-datepicker__input-container input) {
-  width: 100%;
-  padding: 16px;
-  border: 1px solid #D1D5DB;
-  border-radius: 8px;
-  font-size: 16px;
-  font-weight: 400;
-  color: #594A3E;
-  background: white;
-  cursor: pointer;
-}
-
-:global(.react-datepicker__input-container input:focus) {
-  outline: none;
-  border-color: #F5A623;
-}
-
-:global(.react-datepicker__input-container input::placeholder) {
-  color: #ADAEBC;
-}
-
-.buttonSection {
-  display: flex;
-  gap: 16px;
-  justify-content: center;
-  padding: 32px 0;
-}
-
-.cancelButton {
-  width: 442px;
-  height: 60px;
-  background: white;
-  border: 2px solid #D1D5DB;
-  border-radius: 8px;
-  color: #374151;
-  font-size: 16px;
-  font-weight: 400;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.cancelButton:hover {
-  background: #F9FAFB;
-  border-color: #9CA3AF;
-}
-
-.submitButton {
-  width: 438px;
-  height: 60px;
-  background: #F5A623;
-  border: none;
-  border-radius: 8px;
-  color: white;
-  font-size: 16px;
-  font-weight: 400;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.submitButton:hover {
-  background: #E0941A;
-}
-
-/* 반응형 디자인 */
-@media (max-width: 1200px) {
-  .container {
-    padding: 24px 48px;
-  }
-  
-  .formContainer {
-    flex-direction: column;
-  }
-  
-  .progressSection {
-    width: 100%;
-  }
-  
-  .stepsContainer {
-    flex-direction: row;
-    flex-wrap: wrap;
-    gap: 12px;
-  }
-  
-  .stepItem {
-    flex-direction: column;
-    text-align: center;
-    min-width: 80px;
-  }
-}
-
-@media (max-width: 768px) {
-  .container {
-    padding: 16px 24px;
-  }
-  
-  .dateSection {
-    grid-template-columns: 1fr;
-  }
-  
-  .buttonSection {
-    flex-direction: column;
-  }
-  
-  .cancelButton,
-  .submitButton {
-    width: 100%;
-  }
-}
