+++ src/app/advertiser/ads-list/register/styles/FormSections/ParticipationInfoSection.module.css
@@ -1,0 +1,120 @@
+.formSection {
+  background: white;
+  border: 1px solid #E5E7EB;
+  border-radius: 12px;
+  padding: 25px;
+  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.05);
+}
+
+.label {
+  display: block;
+  font-size: 18px;
+  font-weight: 400;
+  color: #594A3E;
+  margin-bottom: 16px;
+}
+
+.required {
+  color: #EF4444;
+  margin-left: 4px;
+}
+
+.dateSection {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 24px;
+}
+
+.dateField {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  width: 300px;
+}
+
+.dateLabel {
+  font-size: 15px;
+  font-weight: 400;
+  color: #594A3E;
+}
+
+.dateInputs {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.dateInputs span {
+  color: #6B7280;
+  font-size: 14px;
+}
+
+.datePicker {
+  flex: 1;
+  padding: 16px;
+  border: 1px solid #D1D5DB;
+  border-radius: 8px;
+  font-size: 16px;
+  font-weight: 400;
+  color: #594A3E;
+  background: white;
+  cursor: pointer;
+}
+
+.datePicker:focus {
+  outline: none;
+  border-color: #F5A623;
+}
+
+/* DatePicker 커스텀 스타일 */
+:global(.react-datepicker-wrapper) {
+  flex: 1;
+}
+
+:global(.react-datepicker__input-container input) {
+  width: 100%;
+  padding: 16px;
+  border: 1px solid #D1D5DB;
+  border-radius: 8px;
+  font-size: 16px;
+  font-weight: 400;
+  color: #594A3E;
+  background: white;
+  cursor: pointer;
+}
+
+:global(.react-datepicker__input-container input:focus) {
+  outline: none;
+  border-color: #F5A623;
+}
+
+:global(.react-datepicker__input-container input::placeholder) {
+  color: #ADAEBC;
+}
+
+@media (max-width: 768px) {
+  .dateSection {
+    grid-template-columns: 1fr;
+  }
+}
+
+.input {
+  width: 100%;
+  padding: 16px;
+  border: 1px solid #D1D5DB;
+  border-radius: 8px;
+  font-size: 16px;
+  font-weight: 400;
+  color: #594A3E;
+  background: white;
+  box-sizing: border-box;
+}
+
+.input:focus {
+  outline: none;
+  border-color: #F5A623;
+}
+
+.input::placeholder {
+  color: #ADAEBC;
+}
